pyproject.toml
[tool.poetry]
name = "rating-map"
version = "0.1.0"
description = ""
authors = ["andreyPromaster <ohotnik17.com@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
psycopg2-binary = "^2.9.3"
djangorestframework-simplejwt = "^5.0.0"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
